refactor DTOs y responsabilidades entre capas, add mapper config

+++ fuentes/BACK/backend/src/main/java/trainer/api/backend/model/dao/IUsuarioRegistroDao.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import trainer.api.backend.model.entity.UsuarioRegistro;
 
-public interface IUsuarioRegistroDao extends CrudRepository<UsuarioRegistro, Integer> {
+public interface IUsuarioRegistroDao extends CrudRepository<UsuarioRegistro, Long> {
 
     @Query("SELECT u FROM UsuarioRegistro u WHERE u.email = ?1")
     UsuarioRegistro findByEmail(String email);
